Move AnimatedPlaceholder to a file
